+++ src/main/kotlin/de/nordgedanken/auto_hotkey/ide/commenter/AhkCommenter.kt
@@ -2,6 +2,10 @@ package de.nordgedanken.auto_hotkey.ide.commenter
 
 import com.intellij.lang.Commenter
 
+/**
+ * Defines the comment character that should be used when the user executes one
+ * of the 'Toggle comment' actions (or invokes its shortcut)
+ */
 class AhkCommenter : Commenter {
     override fun getLineCommentPrefix() = ";"