+++ src/main/kotlin/de/nordgedanken/auto_hotkey/ide/highlighter/AhkSyntaxHighlighter.kt
@@ -6,6 +6,7 @@ import com.intellij.openapi.fileTypes.SyntaxHighlighterBase
 import com.intellij.psi.tree.IElementType
 import de.nordgedanken.auto_hotkey.lang.lexer.AhkLexerAdapter
 import de.nordgedanken.auto_hotkey.lang.psi.AhkTypes.BLOCK_COMMENT
+import de.nordgedanken.auto_hotkey.lang.psi.AhkTypes.DIRECTIVE
 import de.nordgedanken.auto_hotkey.lang.psi.AhkTypes.LINE_COMMENT
 
 class AhkSyntaxHighlighter : SyntaxHighlighterBase() {
@@ -15,9 +16,10 @@ class AhkSyntaxHighlighter : SyntaxHighlighterBase() {
         pack(map(tokenType)?.textAttributesKey)
 
     companion object {
-        fun map(tokenType: IElementType): AhkColor? = when (tokenType) {
-            LINE_COMMENT -> AhkColor.LINE_COMMENT
-            BLOCK_COMMENT -> AhkColor.BLOCK_COMMENT
+        fun map(tokenType: IElementType): AhkHighlighterColor? = when (tokenType) {
+            LINE_COMMENT -> AhkHighlighterColor.LINE_COMMENT
+            BLOCK_COMMENT -> AhkHighlighterColor.BLOCK_COMMENT
+            DIRECTIVE -> AhkHighlighterColor.DIRECTIVE
             else -> null
         }
     }