+++ src/main/kotlin/de/nordgedanken/auto_hotkey/runconfig/core/AhkRunConfigType.kt
@@ -1,26 +1,30 @@
 package de.nordgedanken.auto_hotkey.runconfig.core
 
 import com.intellij.execution.configurations.ConfigurationFactory
-import com.intellij.execution.configurations.ConfigurationType
+import com.intellij.execution.configurations.ConfigurationTypeBase
+import com.intellij.execution.configurations.ConfigurationTypeUtil
 import de.nordgedanken.auto_hotkey.util.AhkConstants
 import de.nordgedanken.auto_hotkey.util.AhkIcons
-import org.jetbrains.annotations.Nls
 
 /**
  * Creates a new type of run config to select when choosing a run config template. Registered in plugin.xml
  */
-class AhkRunConfigType : ConfigurationType {
-    override fun getDisplayName() = AhkConstants.LANGUAGE_NAME
-
-    override fun getConfigurationTypeDescription(): @Nls String {
-        return "AutoHotkey Run Configuration Type"
+class AhkRunConfigType : ConfigurationTypeBase(
+    ID,
+    AhkConstants.LANGUAGE_NAME,
+    DESCRIPTION,
+    AhkIcons.LOGO
+) {
+    init {
+        addFactory(AhkRunConfigFactory(this))
     }
 
-    override fun getIcon() = AhkIcons.FILE
+    val factory: ConfigurationFactory get() = configurationFactories.single()
 
-    override fun getId() = "AHK_RUN_CONFIGURATION"
+    companion object {
+        const val ID = "AhkRunConfiguration"
+        const val DESCRIPTION = "AutoHotkey run configuration"
 
-    override fun getConfigurationFactories(): Array<ConfigurationFactory> {
-        return arrayOf(AhkRunConfigFactory(this))
+        fun getInstance() = ConfigurationTypeUtil.findConfigurationType(AhkRunConfigType::class.java)
     }
 }