+++ src/test/kotlin/de/nordgedanken/auto_hotkey/lang/lexer/AhkLexingTest.kt
@@ -7,6 +7,7 @@ class AhkLexingTest : AhkLexingTestBase() {
 
     override fun createLexer(): Lexer = AhkLexerAdapter()
 
-    fun `test comments`() = doTest()
+    fun `test line comments`() = doTest()
     fun `test block comments`() = doTest()
+    fun `test directives`() = doTest()
 }