+++ src/test/kotlin/de/nordgedanken/auto_hotkey/lang/parser/AhkParsingTestComplete.kt
@@ -1,6 +1,7 @@
 package de.nordgedanken.auto_hotkey.lang.parser
 
 internal class AhkParsingTestComplete : AhkParsingTestBase("complete") {
-    fun `test comments`() = doTest(true)
+    fun `test line comments`() = doTest(true)
     fun `test block comments`() = doTest(true)
+    fun `test directives`() = doTest(true)
 }