+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Syntax highlighting now supports block comments
+- Ide now alerts with a popup notification if there are no Ahk runners configured
+- Right-clicking an Ahk file now offers a 'run' option (only if the file has code)
+- A run icon will appear in the gutter for the first executable line in an Ahk file 
+
+### Changed
+- Modified the Ahk runconfig ID to match the format of the Rust plugin (you will need to modify the "type" in the run config's xml to "AhkRunConfiguration" or just re-create any Ahk run configs you saved) 
+- Changed the AutoHotkey file icon to match the icon in Windows File Explorer
 
 ## [0.4.0] - 2021-02-14
 ### Added